feature(Cleanup) – Restructure file and folder structure
